BACKLOG-9625 Rename a couple of components, extract constants/utils
