+++ src/javascript/components/SiteSettingsSeoApp.constants.js
@@ -1,0 +1,5 @@
+export default {
+    MAPPING_REG_EXP: new RegExp('^/?(?!.*/{2,})[a-zA-Z_0-9\\-\\./]+$'),
+    NB_NEW_MAPPING_ROWS: 1,
+    TABLE_POLLING_INTERVAL: 2000
+};
