+++ src/javascript/components/SiteSettingsSeoEntry.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import {SiteSettingsSeoComponent, SiteSettingsSeoConstants} from './SiteSettingsSeoApp';
+import {SiteSettingsSeoComponent} from './SiteSettingsSeoApp';
 import SiteSettingsSeoWrapper from './SiteSettingsSeoWrapper';
 
 const SiteSettingsSeoEntry = props => <SiteSettingsSeoWrapper component={SiteSettingsSeoComponent} {...props}/>;
 
-export {SiteSettingsSeoEntry, SiteSettingsSeoConstants};
+export {SiteSettingsSeoEntry};
